Moving create test router function to utils, deleting unnecessary ids from home component
