+++ client/src/pages/home/Home.tsx
@@ -1,3 +1,3 @@
 export const Home = () => {
-  return <div data-testid="test-home">home</div>;
+  return <div>home</div>;
 };
